+++ src/app/beer_garden/router.py
@@ -546,8 +546,12 @@ def setup_routing():
     It will then query the database for all local systems and add those to the
     dictionaries as well.
     """
-    for system in db.query(System, filter_params={"local": True}):
-        add_routing_system(system)
+    for system in db.query(
+        System,
+        filter_params={"local": True},
+        include_fields=["instances", "name", "namespace", "version", "garden_name"],
+    ):
+        add_routing_system(system, garden_name=config.get("garden.name"))
 
     # Don't add the local garden
     for garden in get_gardens(include_local=False):
@@ -589,9 +593,18 @@ def add_routing_system(system=None, garden_name=None):
     sure this can handle that without breaking.
 
     """
+    if not garden_name:
+        if not getattr(system, "garden_name", None):
+            for garden in get_gardens():
+                for garden_system in garden.systems:
+                    if garden_system.id == system.id:
+                        garden_name = garden.name
+                        break
+                if garden_name:
+                    break
+        else:
+            garden_name = system.garden_name
     # Default to local garden name
-    garden_name = garden_name or config.get("garden.name")
-
     with routing_lock:
         logger.debug(f"{garden_name}: Adding system {system} ({system.id})")
 
@@ -648,7 +661,7 @@ def add_routing_garden(garden: Garden):
 def handle_event(event):
     """Handle events"""
     if event.name in (Events.SYSTEM_CREATED.name, Events.SYSTEM_UPDATED.name):
-        add_routing_system(system=event.payload, garden_name=event.garden)
+        add_routing_system(system=event.payload)
         return
     elif event.name == Events.SYSTEM_REMOVED.name:
         remove_routing_system(system=event.payload)
@@ -948,16 +961,23 @@ def _system_name_lookup(system: Union[str, System]) -> str:
             version=system.version,
         )
         if len(systems) == 1:
-            for garden in get_gardens():
-                for system in garden.systems:
-                    if systems[0].id == system.id:
-                        with routing_lock:
-                            # Then add routes to systems
-                            add_routing_system(system=system, garden_name=garden.name)
-                        logger.error(
-                            "Router mapping is out of sync, you should consider re-syncing"
-                        )
-                        return garden.name
+            garden_name = systems[0].garden_name
+
+            if not garden_name:
+                for garden in get_gardens():
+                    for system in garden.systems:
+                        if systems[0].id == system.id:
+                            garden_name = garden.name
+                            break
+                    if garden_name:
+                        break
+            with routing_lock:
+                # Then add routes to systems
+                add_routing_system(system=system)
+            logger.error(
+                "Router mapping is out of sync, you should consider re-syncing"
+            )
+            return garden_name
 
     return None