+++ CHANGELOG.md
@@ -12,6 +12,7 @@ TBD
 - Fixed bug in `SystemsList` API using `filter_running` causing SchemaParser to fail(#1889)
 - Removed redundant call to get UtilityService getConfig (#1891)
 - Implemented previous Beer Garden installed version check to determine database migrations required (#1880)
+- Added Support to track Garden Name on System Models
 
 ## 3.30.0rc0